Decompile DRA func_8012D3E8
